requirements.txt
aiohttp==3.7.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
fake-useragent==0.1.11
idna==2.10
lxml==4.6.2
multidict==5.1.0
numpy==1.20.1
pika==1.2.0
PyJWT==1.7.1
python-rucaptcha==3.0.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
vk-api==11.9.1
yarl==1.6.3
psycopg2==2.8.6
gunicorn==20.1.0